requirements.txt
streamlit==1.38.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.1
joblib==1.4.2
plotly==5.24.1
python-dotenv==1.0.1
groq==0.10.0
requests==2.32.3
